requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1;python_version<"3.9"
crispy-bootstrap5==2023.10
Django==4.2.9
django-crispy-forms==2.1
django-forms-dynamic==1.0.0
django-htmx==1.17.2
django-localflavor==4.0
django-widget-tweaks==1.5.0
localflavor==1.9
Pint==0.21.1
python-stdnum==1.19
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
django-dotenv==1.4.2
gunicorn
mysqlclient
whitenoise
pillow
django-import-export
xhtml2pdf
weasyprint
djangorestframework
sentry-sdk
django-crontab
pytest
pytest-django
django-filter
factory_boy
Faker